# Conflicts:
#	Itin/.idea/deploymentTargetDropDown.xml
#	Itin/.idea/misc.xml
#	Itin/app/src/main/AndroidManifest.xml
#	Itin/app/src/main/java/com/example/itin/adapters/DayAdapter.kt

+++ Itin/app/src/main/java/com/example/itin/PreviousTripActivity.kt
@@ -16,11 +16,8 @@ import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.database.DatabaseReference
 import com.google.firebase.database.FirebaseDatabase
 import kotlinx.android.synthetic.main.activity_previous_trip.*
-import java.text.SimpleDateFormat
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
-import java.time.temporal.ChronoUnit
-import java.util.*
 
 class PreviousTripActivity : AppCompatActivity(), PreviousTripAdapter.OnItemClickListener {
 
@@ -205,7 +202,7 @@ class PreviousTripActivity : AppCompatActivity(), PreviousTripAdapter.OnItemClic
                     val cost = it.child(i.toString()).child("Cost").value.toString()
                     val notes = it.child(i.toString()).child("Notes").value.toString()
                     var tripID = it.child(i.toString()).child("TripID").value.toString().toInt()
-                    var activityID = it.child(i.toString()).child("ActivityID").value.toString().toInt()
+                    var activityID = it.child(i.toString()).child("ActivityID").value.toString()
 
                     val activity = Activity(name, time, location, cost, notes, tripID, activityID)
                     actList.add(activity)
